Pipe command line shall warn user on the JWT expiration - code refactoring (fix for #436)
